refactor: desperate with malfunctioning tailwind
