--- .stylelintrc.cjs
@@ -1,3 +1,0 @@
-module.exports = {
-    extends: '@it-incubator/stylelint-config',
-}
