+++ package.json
@@ -29,7 +29,6 @@
     "eslint": "^8.57.1",
     "eslint-config-next": "14.2.15",
     "postcss": "^8.4.47",
-    "stylelint": "^16.10.0",
     "tailwindcss": "^3.4.14",
     "typescript": "^5.6.3"
   }
