+++ src/app/(auth)/create-new-password/page.tsx
@@ -1,3 +1,3 @@
-export default function CreateNewPasswordPage() {
+export function CreateNewPasswordPage() {
   return <h2>Create new password</h2>
 }
