+++ src/app/(auth)/expired-link/page.tsx
@@ -1,3 +1,3 @@
-export default function ExpiredLinkPage() {
+export function ExpiredLinkPage() {
   return <h2>ExpiredLink</h2>
 }
