+++ src/app/(auth)/forgot-password/page.tsx
@@ -1,3 +1,3 @@
-export default function ForgotPasswordPage() {
+export function ForgotPasswordPage() {
   return <h2>Forgot password</h2>
 }
