+++ src/app/(auth)/sign-up/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import Wrapper from '@/components/wrapper/Wrapper'
+import { Wrapper } from '@/components/wrapper/Wrapper'
 import { WtSignUp } from '@/widgets/auth/wt-sign-up'
 
 export default function SignUpPage() {
