+++ src/components/a/index.ts
@@ -1,0 +1,1 @@
+export * from './A'
