+++ src/components/header/AppHeader.tsx
@@ -1,7 +1,7 @@
 import { A } from '@/components/a/A'
 import { Header, LogoComponent } from '@tornata/brooms-ui'
 
-export default function AppHeader() {
+export function AppHeader() {
   return (
     <Header>
       <A href={'/'}>
