+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './a'
+export * from './header'
+export * from './wrapper'
