+++ src/features/auth/index.ts
@@ -1,0 +1,4 @@
+export * from './form-sign-in'
+export * from './form-sign-up'
+export * from './privacy-policy'
+export * from './terms-of-service'
