+++ src/widgets/auth/wt-sign-up.tsx
@@ -1,13 +1,13 @@
 import { A } from '@/components/a/A'
 import { FormSignUp } from '@/features/auth/form-sign-up'
-import { cn } from '@/utils/class-names-merge'
+import { cn } from '@/utils'
 import { Button, Card, Typography } from '@tornata/brooms-ui'
 
 export const WtSignUp = () => {
   return (
     <>
       <Card
-        className={cn('p-6 bg-red-400 flex flex-col items-center justify-center max-w-96 gap-2')}
+        className={cn('p-6 bg-green-400 flex flex-col items-center justify-center max-w-96 gap-2')}
       >
         <Typography asElement={'h2'} variant={'h1'}>
           Sign up
@@ -20,12 +20,8 @@ export const WtSignUp = () => {
           <A href={'/sign-in'}>Sign in</A>
         </Button>
       </Card>
-      <Card className={'bg-red-500 p-6'} style={{ padding: '24px' }}>
-        TEST
-      </Card>
-      <div className={'text-green-300'} style={{ backgroundColor: 'lime' }}>
-        NATIVE DIV
-      </div>
+      <Card className={'bg-amber-600 p-12 m-6'}>TEST</Card>
+      <div className={cn('bg-lime-400 text-red-300 p-16 m-6')}>NATIVE DIV</div>
     </>
   )
 }
